[PR feedback] Dedupe banner helpers into utils; tighten fallback tests
Three functions (resolve_icon_url, download_icon, trim_to_content) were copy-pasted byte-for-byte across generate_rabbit_maximizer.py and generate_rangelink.py, and the trim logic was duplicated again in two test helpers. Moving them to utils.py makes tests call the real production functions instead of copies that could silently drift. download_icon also gets a 10 MB response-size cap for defense-in-depth.

Three fallback tests (title-fallback for rabbit-maximizer and rangelink, tagline-fallback for network-nudge) only checked output dimensions, not rendered content. They now assert against golden images, catching regressions where compose_banner ignores fallback metadata.

Benefits:
- One source of truth for icon download, URL resolution, and trim-to-content
- Fallback tests catch rendering regressions, not just missing output
- download_icon enforces a response-size limit

Ignored Feedback:
- Restrict download_icon to raw.githubusercontent.com URLs: this is a local build/CI script whose icon URL comes from committed repo content. Anyone who can modify the front matter can also modify the Python script directly. The URL allowlist is unnecessary complexity for this threat model.

Ref: https://github.com/couimet/couimet.github.io/pull/136#pullrequestreview-4774967040
